+++ src/core/solution.jl
@@ -35,7 +35,7 @@ function _transform_solution_to_pu!(result, pmitd_data::Dict{String,<:Any}; make
     # per unit specification
     result["solution"]["it"][_PM.pm_it_name]["per_unit"] = pmitd_data["it"][_PM.pm_it_name]["per_unit"]
 
-    # Make mixed units (not per unit)
+    # Make per unit
     _PM.make_per_unit!(result["solution"]["it"][_PM.pm_it_name])
 
     # Transform pmd MATH result to ENG
@@ -47,6 +47,9 @@ function _transform_solution_to_pu!(result, pmitd_data::Dict{String,<:Any}; make
 
     # change PMD dictionary per_unit value (Not done automatically by PMD)
     result["solution"]["it"][_PMD.pmd_it_name]["per_unit"] = true
+
+    # transform pmitd MATH result ref to ENG result ref
+    transform_pmitd_solution_to_eng!(result, pmitd_data)
 
 end
 
@@ -98,6 +101,9 @@ function _transform_solution_to_si!(result, pmitd_data::Dict{String,<:Any}; make
 
     # Transform pmitd solution to PMD SI
     _transform_pmitd_solution_to_si!(result)
+
+    # transform pmitd MATH result ref to ENG result ref
+    transform_pmitd_solution_to_eng!(result, pmitd_data)
 
 end
 
@@ -164,3 +170,76 @@ function sol_data_model!(pmitd::AbstractPowerModelITD, solution::Dict{String,<:A
     _PMD.sol_data_model!(pmd_model, solution["it"][_PMD.pmd_it_name]) # convert distribution system solution.
 
 end
+
+
+"""
+    function transform_pmitd_solution_to_eng!(
+        result::Dict{String,<:Any},
+        pmitd_data::Dict{String,<:Any}
+    )
+
+Transforms the PMITD solution from MATH to ENG model. This transformation
+facilitates the conversion in "pmitd" of buses numbers to buses names according to
+the ENG model. Ex: (100002, 9, 6) -> (100002, voltage_source.3bus_unbal_nogen_mn_2.source, 6)
+"""
+function transform_pmitd_solution_to_eng!(result::Dict{String,<:Any}, pmitd_data::Dict{String,<:Any})
+
+    # get solutions
+    pmitd_sol = result["solution"]["it"][pmitd_it_name]
+
+    # Check for multinetwork and get dictionaries with all buses
+    if haskey(pmitd_sol, "nw")
+        pmitd_sol_data = pmitd_sol["nw"]
+    else
+        pmitd_sol_data = Dict("0" => pmitd_sol)
+    end
+
+    # create empty dict that will replace the math dict in pmitd_sol
+    pmitd_sol_boundary_eng = Dict{String, Any}()
+
+    # loop through nw
+    for (nw, nw_sol_data) in pmitd_sol_data
+
+        # get tran. & dist. buses depending on mn parameter
+        if nw == "0"
+            tran_buses, dist_buses = pmitd_data["it"][_PM.pm_it_name]["bus"], pmitd_data["it"][_PMD.pmd_it_name]["bus"]
+        else
+            tran_buses, dist_buses = pmitd_data["it"][_PM.pm_it_name]["nw"][nw]["bus"], pmitd_data["it"][_PMD.pmd_it_name]["nw"][nw]["bus"]
+        end
+
+        for (boundary_num, boundary_data) in nw_sol_data["boundary"]
+
+            # parse tuple from string
+            boundary_num = eval(Meta.parse(boundary_num))
+
+            # transmission
+            if (haskey(boundary_data, "pbound_fr"))
+                tran_bus = tran_buses[findfirst(x -> boundary_num[2] == x["bus_i"], tran_buses)]
+                tran_bus_name = tran_bus["source_id"][2]
+
+                dist_bus = dist_buses[findfirst(x -> boundary_num[3] == x["bus_i"], dist_buses)]
+                dist_bus_name = dist_bus["source_id"]
+
+                new_name = "("* string(boundary_num[1]) * ", " * string(tran_bus_name) * ", " * dist_bus_name * ")"
+                pmitd_sol_boundary_eng[new_name] = boundary_data
+            end
+
+            # distribution
+            if (haskey(boundary_data, "pbound_to"))
+                dist_bus = dist_buses[findfirst(x -> boundary_num[2] == x["bus_i"], dist_buses)]
+                dist_bus_name = dist_bus["source_id"]
+
+                tran_bus = tran_buses[findfirst(x -> boundary_num[3] == x["bus_i"], tran_buses)]
+                tran_bus_name = tran_bus["source_id"][2]
+
+                new_name = "("* string(boundary_num[1]) * ", " * dist_bus_name * ", " * string(tran_bus_name) * ")"
+                pmitd_sol_boundary_eng[new_name] = boundary_data
+            end
+        end
+
+        # replace ["boundary"] dict in solution
+        nw_sol_data["boundary"] = pmitd_sol_boundary_eng
+    end
+
+end
+