+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Added `autorename.jl` unit tests designed to test errors that must appear when the new naming convention is not followed and `auto_rename=false`.
 - Added documentation to `/docs` that explains in detail the new 'ideal' format for the boundary linking JSON files.
 - Updated documentation guides and README.
+- Added a new transformation function to `solution.jl` called `transform_pmitd_solution_to_eng!` in charge of converting the `pmitd` solution from MATH to ENG model.
 
 ## v0.6.1